microservices/api-metadata-extractor/requirements.txt
fastapi
uvicorn
confluent_kafka
python-multipart
pillow
scikit-learn
joblib
